+++ include/cparsec2/list.h
@@ -42,6 +42,18 @@ extern "C" {
   ELEMENT_TYPE(List(T)) * LIST_END(T)(List(T) xs);                       \
   int LIST_LENGTH(T)(List(T) xs)
 
+#define DEFINE_LIST(T)                                                   \
+  ELEMENT_TYPE(List(T)) * LIST_BEGIN(T)(List(T) xs) {                    \
+    return xs.data;                                                      \
+  }                                                                      \
+  ELEMENT_TYPE(List(T)) * LIST_END(T)(List(T) xs) {                      \
+    return xs.data + xs.len;                                             \
+  }                                                                      \
+  int LIST_LENGTH(T)(List(T) xs) {                                       \
+    return xs.len;                                                       \
+  }                                                                      \
+  END_OF_STATEMENTS
+
 // List(Char)
 #define CharList String
 typedef const char* List(Char);
@@ -97,6 +109,44 @@ FOREACH(DECLARE_LIST, ELEMENT_TYPESET);
   void BUFF_PUSH(T)(Buff(T) * b, ELEMENT_TYPE(Buff(T)) x);               \
   void BUFF_APPEND(T)(Buff(T) * b, List(T) xs);                          \
   List(T) BUFF_FINISH(T)(Buff(T) * b)
+
+#define DEFINE_BUFF(T)                                                   \
+  DEFINE_BUFF_COMMON(T);                                                 \
+  DEFINE_BUFF_FINISH(T)
+
+#define DEFINE_BUFF_COMMON(T)                                            \
+  void BUFF_ENSURE(T)(Buff(T) * b) {                                     \
+    if (!b->data) {                                                      \
+      b->data = mem_malloc(sizeof(ELEMENT_TYPE(Buff(T))) * 8);           \
+      b->capacity = 8;                                                   \
+      b->len = 0;                                                        \
+    }                                                                    \
+    if (b->capacity == b->len) {                                         \
+      b->capacity *= 2;                                                  \
+      b->data = mem_realloc(b->data, sizeof(ELEMENT_TYPE(Buff(T))) *     \
+                                         b->capacity);                   \
+    }                                                                    \
+  }                                                                      \
+  void BUFF_PUSH(T)(Buff(T) * b, ELEMENT_TYPE(Buff(T)) x) {              \
+    BUFF_ENSURE(T)(b);                                                   \
+    b->data[b->len++] = x;                                               \
+  }                                                                      \
+  void BUFF_APPEND(T)(Buff(T) * b, List(T) xs) {                         \
+    ELEMENT_TYPE(List(T))* itr = LIST_BEGIN(T)(xs);                      \
+    ELEMENT_TYPE(List(T))* end = LIST_END(T)(xs);                        \
+    for (; itr != end; itr++) {                                          \
+      BUFF_PUSH(T)(b, *itr);                                             \
+    }                                                                    \
+  }                                                                      \
+  END_OF_STATEMENTS
+
+#define DEFINE_BUFF_FINISH(T)                                            \
+  List(T) BUFF_FINISH(T)(Buff(T) * b) {                                  \
+    List(T) xs = (List(T)){.data = b->data, .len = b->len};              \
+    *b = (Buff(T)){0};                                                   \
+    return xs;                                                           \
+  }                                                                      \
+  END_OF_STATEMENTS
 
 // Buff(Char)
 TYPEDEF_BUFF(Char, char);
